+++ z-paging/uni_modules/z-paging/components/z-paging-empty-view/z-paging-empty-view.vue
@@ -93,7 +93,7 @@
 			emptyViewZIndex: {
 				type: Number,
 				default: function() {
-					return 99
+					return 9
 				}
 			}
 		},